drv/bluetooth_stm32_bluenrg: use random address
This changes the driver to use a random Bluetooth address when
advertising/accepting connections. This will help prevent issues with
Bluetooth stacks caching the attribute database of the official LEGO
firmware and trying to use it for the Pybricks firmware or vice versa.

The Bluetooth chip appears to use the same random address after each
reset, so we don't get the full benefit avoiding caching issues when
the Pybricks firmware GATT database changes.

Issue: https://github.com/pybricks/support/issues/600